requirements.txt
Flask==3.0.3
PyJWT==2.9.0
python-dotenv==1.0.1
Werkzeug==3.0.4
psycopg2-binary==2.9.9
google-genai==0.8.0
